pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]

[project]
name = "prow-jobs-scraper"
description = "Scrape Prow for job results and export them to elasticsearch"
readme = "README.md"
keywords = ["prow"]
license = { file = "LICENSE" }
classifiers = ["Programming Language :: Python :: 3"]
dependencies = [
    "requests==2.27.1",
    "google-cloud-storage==2.3.0",
    "junitparser==2.5.0",
    "pydantic==1.9.1",
    "opensearch-py==1.1.0",
]
dynamic = ["version"]

[project.scripts]
prow-jobs-scraper = "prowjobsscraper.main:main"

[project.optional-dependencies]
test-runner = ["tox==3.25.0"]
unit-tests = ["pytest==7.1.2", "pytest-cov==3.0.0", "pytest-httpserver==1.0.4"]
coverage = ["coverage[toml]==6.3.3"]
lint = ["black==22.3.0", "isort==5.10.1", "mypy==0.950"]

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = clean,python3.9,report

[testenv]
usedevelop=True
commands = pytest --cov=src --cov-append --cov-report=term-missing --junitxml=junit-{envname}.xml --junit-prefix={envname} {posargs}
deps = .[unit-tests]
depends =
    {python3.9}: clean
    report: python3.9

[testenv:report]
deps = .[coverage]
skip_install = true
commands =
    coverage report
    coverage html

[testenv:clean]
deps = .[coverage]
skip_install = true
commands = coverage erase
"""
